pyproject.toml
[tool.poetry]
name = "hi_en_sentiment"
version = "0.1.0"
description = ""
authors = ["chris schaller <chris.schaller@posteo.de>", "pavlos musenidis <>"]

[tool.poetry.dependencies]
python = "^3.8"
sklearn = "^0.0"
pandas = "^1.0.3"
tensorflow = "^2.2.0"
scikit-learn = "^0.23.0"
scipy = "^1.4.1"
tensorflow_hub = "^0.8.0"
bert-for-tf2 = "^0.14.4"
tqdm = "^4.46.1"
numpy = "^1.18.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.14.0"
ipykernel = "^5.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
